+++ src/main/java/pro/sky/telegrambot/utils/Patterns.java
@@ -2,17 +2,17 @@ package pro.sky.telegrambot.utils;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.request.SendMessage;
+import lombok.experimental.UtilityClass;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.regex.Pattern;
 
+@UtilityClass
 public class Patterns {
 
     public static final Pattern PATTERN = Pattern.compile("(\\d{2}\\.\\d{2}\\.\\d{4}\\s\\d{2}:\\d{2})(\\s+)(.+)");
-
-    public static final LocalDateTime CURRENT_TIME_TO_MINUTES = LocalDateTime.now().truncatedTo(ChronoUnit.MINUTES);
 
     public static final DateTimeFormatter TIME_FORMATTER = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm");
