+++ src/main/resources/application.properties
@@ -2,6 +2,7 @@ spring.datasource.url = jdbc:postgresql://localhost:5432/telegramBot
 spring.datasource.username = Vladimir
 spring.datasource.password = 12345
 telegram.bot.token = 7341978843:AAG63bw6kcpTqTmUJOZaX2LfisuQ7Xk0UyE
-spring.jpa.hibernate.ddl-auto = update
+spring.jpa.hibernate.ddl-auto = validate
+spring.liquibase.enabled = true
 
 spring.liquibase.change-log = classpath:/liquibase/changelog-master.yaml
