+++ src/main/resources/liquibase/scripts/scripts.sql
@@ -1,0 +1,1 @@
+Empty file.
